botty/requirements.txt
discord.py==2.2.2
Flask==2.2.3
python-dotenv==1.0.0
requests==2.25.1
